Change queryTable doc to use result.entries
This threw me for a loop for about 30 minutes. The documentation says result.entitites will contain the entities returned from the query, but the code (line 849) uses result.entries.

http://azure.github.io/azure-storage-node/TableService.html#queryEntities

It might make more sense for everything to be entities, but unless both entities and entries are in the object, that would be a breaking change.